Solving python import references and changing python 2.X syntax to 3.X syntax
Some examples:
1) print()
2) if key in dict
3) from . import
4) configparser, Raw
5) logger.warning()
6) Exception as msg

+++ plugins/et2216e/et2216e/stepentry.py
@@ -7,7 +7,7 @@ import core.stepentry
 
 def parseFromNode(node):
     stepentry = UP150StepEntry()
-    temperature = long(node.getChildNamed('temperature').getValue())
+    temperature = int(node.getChildNamed('temperature').getValue())
     stepentry.setTemperature(temperature)
     return stepentry
